refactor(class): 강의 도메인 필드 타입 수정 - Int -> Long으로 수정

+++ src/main/kotlin/com/liveklass/domain/DomainException.kt
@@ -13,6 +13,6 @@ class ClassEnrollmentException(val classId: Long, val currentStatus: ClassStatus
 
 class ClassCapacityExceededException(
     val classId: Long,
-    val capacity: Int,
-    val enrolledCount: Int
+    val capacity: Long,
+    val enrolledCount: Long
 ) : DomainException("강의 정원이 초과되었습니다. classId=$classId, capacity=$capacity, enrolledCount=$enrolledCount")
